// frontend/src/api/types/question.ts
/**
 * Generated by orval v8.0.0-rc.2 🍺
 * Do not edit manually.
 * Education platform
 * OpenAPI spec version: 0.0.1
 */
import type { QuestionType } from './questionType';
import type { Answer } from './answer';

export interface Question {
  /** Уникальный идентификатор вопроса */
  id: string;
  /** Текст вопроса */
  text: string;
  /** Тип вопроса */
  type: QuestionType;
  /** ID теста, к которому относится вопрос */
  quizId: string;
  /** Список ответов на вопрос */
  answers: Answer[];
  /** Дата создания записи */
  createdAt: string;
  /** Дата последнего обновления записи */
  updatedAt: string;
}