// frontend/src/api/types/questionType.ts
/**
 * Generated by orval v8.0.0-rc.2 🍺
 * Do not edit manually.
 * Education platform
 * OpenAPI spec version: 0.0.1
 */

/**
 * Тип вопроса
 */
export type QuestionType = typeof QuestionType[keyof typeof QuestionType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const QuestionType = {
  SINGLE_CHOICE: 'SINGLE_CHOICE',
  MULTIPLE_CHOICE: 'MULTIPLE_CHOICE',
} as const;